refactor: 使用 `NameResolver` 优化处理注释的逻辑
