+++ src/Processor/ClassProcessor.php
@@ -12,7 +12,6 @@ use PhpAccessor\File\File;
 use PhpAccessor\Processor\Builder\DataBuilder;
 use PhpAccessor\Processor\Method\AccessorMethod;
 use PhpParser\BuilderFactory;
-use PhpParser\Comment\Doc;
 use PhpParser\Node;
 use PhpParser\Node\Attribute;
 use PhpParser\Node\Expr\Include_;
@@ -22,7 +21,6 @@ use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Expression;
 use PhpParser\Node\Stmt\Namespace_;
 use PhpParser\Node\Stmt\Property;
-use PhpParser\Node\Stmt\Use_;
 use PhpParser\NodeFinder;
 use PhpParser\NodeVisitorAbstract;
 
@@ -35,29 +33,6 @@ class ClassProcessor extends NodeVisitorAbstract
 
     private string $classname = '';
 
-    private string $namespace = '';
-
-    /**
-     * @var array<string, string>
-     */
-    private array $usedClassNames = [];
-
-    private const PRIMITIVE_TYPES = [
-        'bool' => 'bool',
-        'boolean' => 'bool',
-        'string' => 'string',
-        'int' => 'int',
-        'integer' => 'int',
-        'float' => 'float',
-        'double' => 'float',
-        'array' => 'array',
-        'object' => 'object',
-        'callable' => 'callable',
-        'resource' => 'resource',
-        'mixed' => 'mixed',
-        'iterable' => 'iterable',
-    ];
-
     private bool $genCompleted = false;
 
     private TraitAccessor $traitAccessor;
@@ -65,18 +40,13 @@ class ClassProcessor extends NodeVisitorAbstract
     private NodeFinder $nodeFinder;
 
     public function __construct(
-        private bool $genMethod
+        private bool $genMethod,
     ) {
         $this->nodeFinder = new NodeFinder();
     }
 
     public function enterNode(Node $node)
     {
-        if ($node instanceof Use_) {
-            $this->collectUsedClassNames($node);
-
-            return null;
-        }
         if (!$node instanceof Class_ || empty($node->attrGroups)) {
             return null;
         }
@@ -101,9 +71,6 @@ class ClassProcessor extends NodeVisitorAbstract
         if (empty($properties)) {
             return null;
         }
-
-        $this->processPropertiesDocComment($properties);
-
         $this->generateAllMethods($node->namespacedName->toString(), $properties, $attributeProcessor);
         /** @var ClassMethod[] $originalClassMethods */
         $originalClassMethods = $this->nodeFinder->findInstanceOf($node, ClassMethod::class);
@@ -202,63 +169,5 @@ class ClassProcessor extends NodeVisitorAbstract
     public function getTraitAccessor(): TraitAccessor
     {
         return $this->traitAccessor;
-    }
-
-    private function collectUsedClassNames(Use_ $node): void
-    {
-        foreach ($node->uses as $use) {
-            $aliasName = $use->name->getLast();
-            if (!empty($use->alias)) {
-                $aliasName = $use->alias->toString();
-            }
-            $this->usedClassNames[strtolower($aliasName)] = $use->name->toString();
-        }
-    }
-
-    /**
-     * @param Property[] $properties
-     */
-    private function processPropertiesDocComment(array $properties): void
-    {
-        foreach ($properties as $property) {
-            if (empty($docComment = $property->getDocComment())) {
-                continue;
-            }
-
-            if (!preg_match('/(?<=@var\s)[^\s]+/', $docComment->getText(), $matches)) {
-                continue;
-            }
-
-            $type = $matches[0];
-            $isArray = str_ends_with($type, '[]');
-            if ($isArray) {
-                $type = substr($type, 0, -2);
-            }
-
-            if (isset(self::PRIMITIVE_TYPES[$type])) {
-                continue;
-            }
-
-            $type = $this->getUsedClassName($type) . ($isArray ? '[]' : '');
-            $text = str_replace($matches[0], $type, $docComment->getText());
-
-            $property->setDocComment(new Doc($text));
-        }
-    }
-
-    private function getUsedClassName(string $type): ?string
-    {
-        $alias = ($pos = strpos($type, '\\')) === false ? $type : substr($type, 0, $pos);
-        $loweredAlias = strtolower($alias);
-
-        if (isset($this->usedClassNames[$loweredAlias])) {
-            if (false !== $pos) {
-                return $this->usedClassNames[$loweredAlias] . substr($type, $pos);
-            }
-
-            return $this->usedClassNames[$loweredAlias];
-        }
-
-        return "{$this->namespace}\\{$type}";
     }
 }
